Trying to using tex obj generating instead of wall obj (dynamic textures)

+++ include/pixelopolis/draw_builder_texture.h
@@ -33,7 +33,7 @@ struct TexPartObj {
 
 struct TileInfo {
     struct TexObj* obj;
-    uint8_t depth;
+    uint32_t depth;
     uint32_t width;
 };
